requirements.txt
annotated-types==0.7.0
anyio==4.11.0
boto3==1.40.49
botocore==1.40.49
click==8.3.0
colorama==0.4.6
fastapi==0.118.3
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.10
jmespath==1.0.1
psycopg2-binary==2.9.11
pydantic==2.12.0
pydantic_core==2.41.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
